+++ conf/urls.py
@@ -663,6 +663,20 @@ international_redirects_urls = [
     ),
 ]
 
+export_vouchers_urls = [
+    url(
+        r'^export-vouchers/$',
+        contact.views.ExportVoucherFormView.as_view(),
+        name='export-voucher-form'
+    ),
+    url(
+        r'^export-vouchers/sent/$',
+        contact.views.ExportVoucherSuccessView.as_view(),
+        name='export-voucher-success'
+    )
+]
+
+
 urlpatterns += legacy_urls
 urlpatterns += deprecated_urls
 urlpatterns += euexit_urls
@@ -675,3 +689,6 @@ urlpatterns += community_urls
 urlpatterns += ukef_urls
 urlpatterns += marketing_urls
 urlpatterns += international_redirects_urls
+
+if settings.FEATURE_FLAGS['EXPORT_VOUCHERS_ON']:
+    urlpatterns += export_vouchers_urls